+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/OracleDialect.cs
@@ -2,5 +2,7 @@ namespace SqlArtisan;
 
 internal sealed class OracleDialect : IDbmsDialect
 {
-    public string GetParameterName(int index) => $":{index}";
+    public string AliasQuote => "\"";
+
+    public string ParameterMarker => ":";
 }