+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - **New Rule:** Remove underscores from the SQL keyword and convert to PascalCase.
   - **Example:** `SysTimestamp` has been renamed to `Systimestamp`, and `LPad` has been renamed to `Lpad`.
 ### Fixed
+- Corrected table alias quoting for MySQL. (#31)
 - **SqlArtisan.TableClassGen**: Fixed the generated base class to be `DbTableBase` instead of the incorrect `AbstractTable`. (#29)
 
 ## [0.1.0-alpha.17] - 2025-06-16